fix(examples): tls example generate certificate and key with rcgen
