--- examples/tls/.gitignore
@@ -1,6 +1,0 @@
-/target
-/Cargo.lock
-*.data*
-*perf*
-*linux-perf*
-*flamegraph*
