--- examples/tls/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "tls2"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-tokio = "1.32.0"
-tokio-rustls = "0.24.1"
-yahf = { path = "../.." }
