--- examples/tls/src/tls_config/local.cert
@@ -1,22 +1,0 @@
------BEGIN CERTIFICATE-----
-MIIDszCCApugAwIBAgIUbiD2u3gGxeU1P+kchD+XAVDN+IIwDQYJKoZIhvcNAQEL
-BQAwaTELMAkGA1UEBhMCQlIxETAPBgNVBAgMCEdvacODwqFzMRMwEQYDVQQHDApH
-b2nDg8KibmlhMQ0wCwYDVQQKDARZQUhGMRQwEgYDVQQLDAtEZXZlbG9wbWVudDEN
-MAsGA1UEAwwEWUFIRjAeFw0yMzA4MjIxNDIyMjFaFw0yODA4MjAxNDIyMjFaMGkx
-CzAJBgNVBAYTAkJSMREwDwYDVQQIDAhHb2nDg8KhczETMBEGA1UEBwwKR29pw4PC
-om5pYTENMAsGA1UECgwEWUFIRjEUMBIGA1UECwwLRGV2ZWxvcG1lbnQxDTALBgNV
-BAMMBFlBSEYwggEiMA0GCSqGSIb3DQEBAQUAA4IBDwAwggEKAoIBAQDRiZ/P2tUN
-vAWvfQRcWeHJtAXVDqYBcCJJue2AanP2uChzsQ9fiD7aXuvGaSpsH959ntq7TyxV
-2hxzPhgrpBm8YvoKBm1b8q5dZRAuchWIzLzgYP2kD2MLX34CmHfGeJhfws613/RU
-+acGEdxobI4NZxWwyt/CwMi0XPmWyJSCQhoG29v7apd2GBF0tsQCV9J4eb+RH90v
-1b5CgYL5k3wxC+ZL0HA1Te5+mrIFYSlXlgOb/UN+q6TEvzP26ym90jmBjRNvVIgh
-CMeChj5h8x4IFsUFtoh752SFQdDEWgecLw5r/oIe3FLNV6rTv8lmZ9dSMTbdZcvJ
-W9gZLjtcrYwrAgMBAAGjUzBRMB0GA1UdDgQWBBSaywPsy7+Ye3kUE/FAcmFV0P3t
-MDAfBgNVHSMEGDAWgBSaywPsy7+Ye3kUE/FAcmFV0P3tMDAPBgNVHRMBAf8EBTAD
-AQH/MA0GCSqGSIb3DQEBCwUAA4IBAQBNifZfftNJQpmgnE2AV6lWQuvlr5LVWHmF
-6p+S1f5xi669nJb4tk4TPNjr7beTNqsdbWvK+a9TrsQKBilXlzX7PYTdsmyQKT7Z
-JH2gl9eYlePguZK5BwM5V0UHyM7deg3mcz740JDH/UHEAgVGpu69Eh6o0LSjX6Xz
-iDrtNAjUrBoZFRZl+0qa5xG4GwN5fy4O+tNM6JKVkAywFKSMDoUM8PRW/Jf8OFDw
-Pzo2RdpDrHw/Y26KYYYfyMUGFkjaN29VYO8/RWOJK+1diKrLunAj9jtXUuFZkZjC
-rET5H8r6Wy/hzlzVwZUSJ/JDcbneHvn4EAqEIaKiQfoctYXAYPNQ
------END CERTIFICATE-----
