--- examples/tls/src/tls_config/local.key
@@ -1,30 +1,0 @@
------BEGIN ENCRYPTED PRIVATE KEY-----
-MIIFHDBOBgkqhkiG9w0BBQ0wQTApBgkqhkiG9w0BBQwwHAQIkKkKtvIlJc0CAggA
-MAwGCCqGSIb3DQIJBQAwFAYIKoZIhvcNAwcECAw7ztD+4BIIBIIEyKnCFtJIy41o
-Zkn6dPgOF2fw0zSpK+eHcAwPHyIFrBNtDinUELiQ7suEhYipN3iHd2Vddzryixuz
-CpLGWNn5L3eZVfJK9dp/wDzNt3Y+NDnFXTRBL5SzXgjhJpyEzZlG/uoK1LIp1ZFF
-Ze6QyQsgxTeOGVb+/QEIN9b21fwiy2chnT+Z+urcexGSMOmJy/Frnh8hdi3MyD8E
-5vZBHosYlNB/Y6l1J/no7fs7XH+HuTeU7Hrh7Md+p0ICe+BICezn2r9pDYs0L2vB
-7wZQnmWmbfIQWtXUknK8QgaQSJMlzgH/yr+DJIOOuh464i+AROXuiDpWe7t/1I9L
-n8GmHxV4ZNJ77F/ITUSELIgnlfhdwWtKY+f1vV9PX17W0VeRgoXlILkfzG+G3e4c
-uHBIoFd1oLu9zv9np80vgmV01W+3MH5kSJm7YuHhoRDr204r0uJK0SHjIZXVyRq0
-0omRusBzR9ZCZ4m1yymqDCZxRslMeZRBPBQOwnQoiuwXeoddIp+8KqCGdtfV2S6/
-+dpbFPE1GTGkqrYY7eJA22Mp0OdrxexKquwwZncvXQSrajkVn8Js/H4N7JApL9SH
-O8W+L7J/g7jIupJi2ZvJJIUakdhwzFPHFRsqstpKEB9H+DCrNzf+7/NtvzRD999y
-J2eoJgZT3KIGJ+b+Y35G+GSf0HNA+St46/WypccssH+GeQIaaxkdTe+rNBEpbJSM
-kjB2j9TJ4SKSJzfrn8PUq8LdxAuGEdKLm/taBuMAWi6bKDquqZqxAvmL1W6M13z2
-qbP4ukD8rb3Uzy9bsj+7O2FTkhwLxrudbeje6QOB+JY4/h1P3vwWxke4RcmA0C7Y
-NqlczOBV8e3E4M2mKPp5Jmv/2EHP6k0D55tTKuBj9P65dRW+lPYrUsm+RAtdwuCn
-v8kLMCmWMwU/PJ1qjfOvKmNAVJTvC6AfSTr9bX0vR5nsaFga0W3WIAqUI0GFAsfS
-7//5Uu7mhgcD5UvVeuHeJPiM8Cju+HA37OIozqwCHHsXHKl9irY9pyVdXW8Zzyim
-M3Ddmf89DzRBwmxXnU/yFh+jZA57JPiV6K1CIoHLMcGyxDYZEWPzziO2lcWuSj8h
-hMNJfIchKPNWXc6m2zqxYkf9rkrV5PtEKCJZio/c22RVk6U5/EjlLVF6oJ99COYO
-K60cTcvNU05N7k0Bok2OsPaWQtUC8OpMZFcqXIvQ4wrqFOAgR6a8NfWfDxqbrsyt
-hG7zmyChrHUHNxEGoRr+WSSZM6eq3L4LAjXr5XvPZku9e7sX/YTH9jd7staS65N4
-AXickgO812qv4VbbF9WqVSsFeN0O3iag2GZA4Lmlf5OUZ3L0AlD4Jlu2o86Q5BAV
-p8rD+UYMltOgUVIV9xK+rRD7TcxwyUAVYVsTN0JajLD5pDqmZQ4vjzDfU8NUhCZV
-7xNxosL6dwP8uJzNTIHYNoVQSyF7ldaisvqSDCPRO7pC9ZIDJ56m4fnBpo6BJaC2
-CNYomtY8cJsWw9W5Bu1NfkQ/zZj8LShggHksgtQcUgJr5zdTlKjCAsfyh4768Laa
-dPcX7muhFvHTCaEWlJtA/yJknOULDISqgzwq56SypeEZnQxStuHFHykGlruaBcKk
-eDIt0HX8IEJs4Hcz3USNoQ==
------END ENCRYPTED PRIVATE KEY-----
